refactor mgfn modules
