+++ src/loss/__init__.py
@@ -1,0 +1,6 @@
+from .base import (  # noqa # pylint: disable=unused-import
+    TemporalSmoothnessLoss,
+    SparsityLoss,
+    ContrastiveLoss,
+)
+from .mgfn import MGFNLoss # noqa # pylint: disable=unused-import
