+++ src/models/mgfn/modeling_mgfn.py
@@ -11,7 +11,7 @@ from transformers import PreTrainedModel
 from transformers.file_utils import ModelOutput
 
 from .configuration_mgfn import MGFNConfig
-from .loss_mgfn import (
+from ...loss import (
     TemporalSmoothnessLoss,
     SparsityLoss,
     MGFNLoss,
